Address warnings

+++ src/aniso_weights.c
@@ -46,7 +46,7 @@ void initialize_weights_AnIso(int nodes, double *zeta, double Lv, double lam, in
   FILE *fidWeights;
   char buffer_weights[100];
   char output_buffer[100];
-  size_t readFlag;
+  int readFlag;
 
   MPI_Comm_size(MPI_COMM_WORLD,&numNodes);
   MPI_Comm_rank(MPI_COMM_WORLD,&rank);
@@ -72,7 +72,7 @@ void initialize_weights_AnIso(int nodes, double *zeta, double Lv, double lam, in
     if((fidWeights = fopen(buffer_weights,"r"))) {
       printf("Loading weights from file %s\n",buffer_weights);
       for(i=0;i<N*N*N;i++) { 
-	readFlag = fread(conv_weights[i],sizeof(double),N*N*N,fidWeights);
+	readFlag = (int) fread(conv_weights[i],sizeof(double),N*N*N,fidWeights);
 	if(readFlag != N*N*N) {
 	  printf("Error reading weight file\n");
 	  exit(1);
